+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.19.1 / 2021-10-25
+
+-   add Table Pagination - Data Table Test
+-   need Fix in Release:
+      TAblePagination.js line 69 - add "this.setPageCount();"
+
 # 0.18.1 / 2021-10-21
 
 -   Add Mocha Test